+++ python/code_challenges/tree_breadth_first/README.md
@@ -1,0 +1,26 @@
+# Daily Code Challenge
+
+## Breadth-first Traversal.
+Author: Odeh Abuzaid
+
+---
+
+### Problem Domain
+Given a Binary Tree , Traverse the input tree using a Breadth-first approach and return list of nodes values in the order they were encountered
+
+---
+### Big O
+
+
+| Time | Space |
+| :----------- | :----------- |
+| O(n)  | O(1) |
+
+---
+
+## Whiteboard Process
+<img src="./Capture.JPG" alt="">
+
+## Approach & Efficiency
+walk the tree nodes using Breadth-first Algorithm with a Queue ,after each dequeue will yield the nodes values
+
