+++ python/code_challenges/tree_breadth_first/pyproject.toml
@@ -1,0 +1,15 @@
+[tool.poetry]
+name = "tree_breadth_first"
+version = "0.1.0"
+description = ""
+authors = ["Odeh <abuzaid163@hotmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.9"
+
+[tool.poetry.dev-dependencies]
+pytest = "^5.2"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
