+++ python/code_challenges/tree_breadth_first/tree_breadth_first/__init__.py
@@ -1,0 +1,1 @@
+__version__ = '0.1.0'
