+++ python/code_challenges/tree_breadth_first/tree_breadth_first/tree_bf.py
@@ -1,7 +1,4 @@
 
-from tree import BTree, Tree
-
-
 class Queue:
   def __init__(self, collection=[]):
     self.data = collection
@@ -17,7 +14,7 @@ class Queue:
   def dequeue(self):
     return self.data.pop(0)
 
-def tree_breadth_first(tree):
+def tree_bfs(tree):
     """
     function which returns a list of items in a given tree using bfs algorithm
 
@@ -29,34 +26,12 @@ def tree_breadth_first(tree):
     breadth = Queue()
     breadth.enqueue(tree.root)
 
-    list_of_items = []
-
     while breadth.peek():
       front = breadth.dequeue()
-      list_of_items += [front.data]
+      yield front.data
 
       if front.left:
         breadth.enqueue(front.left)
 
       if front.right:
         breadth.enqueue(front.right)
-
-    return list_of_items
-
-
-
-def test():
-    # Arrange
-    tree_ = Tree()
-    BTree.add(tree_,1)
-    BTree.add(tree_,2)
-    BTree.add(tree_,3)
-    BTree.add(tree_,4)
-    # Actual
-    actual = tree_breadth_first(tree_)
-    # Expected
-    excepted = [1, 2, 3, 4]
-    # Assert
-    print('Pass') if actual == excepted else print('Fail')
-
-test()
